fix: create deprecated files and imports for all the other tokens
