+++ packages/forma-36-tokens/src/tokens/colors/colors-contrast.js
@@ -1,10 +1,3 @@
-const colorsBlack = require('./colors-black');
-const colorsGray = require('./colors-gray');
-
-const colorsContrast = {
-  'color-contrast-dark': colorsBlack['color-black'],
-  'color-contrast-mid': colorsGray['gray-900'],
-  'color-contrast-light': colorsGray['gray-800'],
-};
+const colorsContrast = require('./colors-contrast-deprecated');
 
 module.exports = colorsContrast;
