+++ packages/forma-36-tokens/src/tokens/colors/colors-coral-deprecated.js
@@ -1,0 +1,8 @@
+const colorsRed = require('./colors-red');
+
+const colorsCoral = {
+  'color-coral-dark': colorsRed['red-200'],
+  'color-coral-mid': colorsRed['red-100'],
+};
+
+module.exports = colorsCoral;
