+++ packages/forma-36-tokens/src/tokens/colors/colors-coral.js
@@ -1,8 +1,3 @@
-const colorsRed = require('./colors-red');
-
-const colorsCoral = {
-  'color-coral-dark': colorsRed['red-200'],
-  'color-coral-mid': colorsRed['red-100'],
-};
+const colorsCoral = require('./colors-coral-deprecated');
 
 module.exports = colorsCoral;
