+++ packages/forma-36-tokens/src/tokens/colors/colors-elements.js
@@ -1,11 +1,3 @@
-const colorsGray = require('./colors-gray');
-
-const colorsElements = {
-  'color-element-darkest': colorsGray['gray-500'],
-  'color-element-dark': colorsGray['gray-400'],
-  'color-element-mid': colorsGray['gray-300'],
-  'color-element-light': colorsGray['gray-200'],
-  'color-element-lightest': colorsGray['gray-100'],
-};
+const colorsElements = require('./colors-elements-deprecated');
 
 module.exports = colorsElements;
