+++ packages/forma-36-tokens/src/tokens/colors/colors-ice.js
@@ -1,8 +1,3 @@
-const colorsBlue = require('./colors-blue');
-
-const colorsIce = {
-  'color-ice-dark': colorsBlue['blue-200'],
-  'color-ice-mid': colorsBlue['blue-100'],
-};
+const colorsIce = require('./colors-ice-deprecated');
 
 module.exports = colorsIce;
