+++ packages/forma-36-tokens/src/tokens/colors/colors-mint.js
@@ -1,8 +1,3 @@
-const colorsGreen = require('./colors-green');
-
-const colorsMint = {
-  'color-mint-dark': colorsGreen['green-200'],
-  'color-mint-mid': colorsGreen['green-100'],
-};
+const colorsMint = require('./colors-mint-deprecated');
 
 module.exports = colorsMint;
