+++ packages/forma-36-tokens/src/tokens/colors/colors-text-deprecated.js
@@ -1,0 +1,11 @@
+const colorsGray = require('./colors-gray');
+
+const colorsText = {
+  'color-text-dark': colorsGray['gray-900'],
+  'color-text-base': colorsGray['gray-800'],
+  'color-text-mid': colorsGray['gray-700'],
+  'color-text-light': colorsGray['gray-600'],
+  'color-text-lightest': colorsGray['gray-500'],
+};
+
+module.exports = colorsText;
