+++ packages/forma-36-tokens/src/tokens/colors/colors-text.js
@@ -1,11 +1,3 @@
-const colorsGray = require('./colors-gray');
-
-const colorsText = {
-  'color-text-dark': colorsGray['gray-900'],
-  'color-text-base': colorsGray['gray-800'],
-  'color-text-mid': colorsGray['gray-700'],
-  'color-text-light': colorsGray['gray-600'],
-  'color-text-lightest': colorsGray['gray-500'],
-};
+const colorsText = require('./colors-text-deprecated');
 
 module.exports = colorsText;
