+++ packages/forma-36-tokens/src/tokens/colors/scale-gray.js
@@ -1,0 +1,13 @@
+const scaleGray = {
+  'gray-100': '#F7F9FA',
+  'gray-200': '#E7EBEE',
+  'gray-300': '#CFD9E0',
+  'gray-400': '#AEC1CC',
+  'gray-500': '#67728A',
+  'gray-600': '#5A657C',
+  'gray-700': '#414D63',
+  'gray-800': '#1B273A',
+  'gray-900': '#111B2B',
+};
+
+module.exports = scaleGray;
